requirements.txt
streamlit>=1.36.0
pandas>=2.1.0
pycaret>=3.3.0
openai>=1.0.0
python-dotenv>=1.0.0
langfuse==2.6.0
boto3>=1.26.0
scikit-learn>=1.3.0
